requirements.txt
click==8.1.3
colorama==0.4.4
DateTime==4.4
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.3.0
pytz==2022.1
SQLAlchemy==1.4.36
uuid==1.30
Werkzeug==2.1.2
zope.interface==5.4.0
Babel==2.10.1
bcrypt==3.2.2
blinker==1.4
cffi==1.15.0
click==8.1.3
colorama==0.4.4
DateTime==4.4
Flask==2.1.2
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.4.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-User==0.6.21
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pytz==2022.1
speaklater==1.3
SQLAlchemy==1.4.36
uuid==1.30
Werkzeug==2.1.2
WTForms==3.0.1
zope.interface==5.4.0
